Fix rust-orm decoding numeric price column as f64 instead of Decimal
store_fruit_price::Model::price was typed f64 against a numeric(12,2)
column. sqlx's Decode<Postgres> for f64 only accepts the FLOAT8 OID, not
NUMERIC, so every SeaORM query touching price failed at runtime with a
type-mismatch error - surfacing as 500s from rust-orm and, on the list
endpoint, an empty response body.

The raw-SQL path avoids this by casting price::float8 in the query text;
SeaORM has no per-query cast for entity-derived queries, so the field is
typed Decimal (via sea-orm's with-rust_decimal feature) and narrowed to
f64 in repository_orm.rs when building the DTO, the same numeric-to-f64
narrowing done in Rust instead of in SQL.

+++ rust/src/entities/store_fruit_price.rs
@@ -12,7 +12,9 @@ pub struct Model {
     pub fruit_id: i64,
     #[sea_orm(primary_key, auto_increment = false, column_name = "store_id")]
     pub store_id: i64,
-    pub price: f64,
+    // `numeric(12,2)` in the DB - `f64` would fail to decode at runtime (sqlx's `Decode<Postgres>
+    // for f64` only accepts the FLOAT8 OID). Converted to `f64` for the DTO in repository_orm.rs.
+    pub price: Decimal,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
